Refactor: Split off scalar operations

+++ src/main/list_operations.cc
@@ -974,7 +974,7 @@ const ops_table_entry ops_table[] = {
 };
 
 int
-add_list_op(as_operations* ops, int64_t opcode, Local<Object> op, LogInfo* log)
+add_list_op(as_operations* ops, uint32_t opcode, Local<Object> op, LogInfo* log)
 {
 	opcode = opcode ^ LIST_OPS_OFFSET;
 	const ops_table_entry *entry = &ops_table[opcode];
